requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.16.1
flask==3.0.3
fonttools==4.55.0
fsspec==2024.10.0
huggingface-hub==0.26.2
idna==3.10
importlib-metadata==8.5.0
importlib-resources==6.4.5
inquirerpy==0.3.4
itsdangerous==2.2.0
jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
networkx==3.1
numpy==1.24.4
packaging==24.2
pfzy==0.3.4
pillow==10.4.0
prompt-toolkit==3.0.48
pyparsing==3.1.4
pytesseract==0.3.13
python-dateutil==2.9.0.post0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.4.5
six==1.16.0
sympy==1.13.3
tokenizers==0.20.3
torch==2.4.1
tqdm==4.67.0
transformers==4.46.3
typing-extensions==4.12.2
urllib3==2.2.3
wcwidth==0.2.13
werkzeug==3.0.6
zipp==3.20.2
